services/infrastructure-generation/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
jinja2==3.1.2
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
hcl2==0.3.20
python-terraform==0.10.1
aws-cdk-lib==2.108.0
constructs==10.3.0
jsii==1.91.0